Navigation improvements

+++ src/app/components/side-bars/common-tools/delete-story.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core'
-import { User, StoryDocument, StoryMetaData } from './../../../_models'
-import { StoryEditorService, StoryService, AuthenticationService, AlertService } from './../../../_services'
+import { User, StoryMetaData } from './../../../_models'
+import { StoryService, AuthenticationService, AlertService, UIService, DynamicForm } from './../../../_services'
 import { Router } from '@angular/router';
 
 @Component({
@@ -15,6 +15,8 @@ export class DeleteStoryComponent implements OnInit {
   private enabled: boolean
 
   constructor(
+    private router: Router,
+    private uiService: UIService,
     private storyService: StoryService,
     private authenticationService: AuthenticationService,
     private alertService: AlertService
@@ -40,11 +42,21 @@ export class DeleteStoryComponent implements OnInit {
   deleteStory() {
     if (this.enabled) {
       this.authenticationService.withLoggedInUser().then((user: User) => {
-        this.storyService.deleteStory(this.storyId).then(() => {
-          this.alertService.success("Story Deleted!")
-        }).catch(e => {
-          this.alertService.error(e)
+        let form: DynamicForm = new DynamicForm("Create Story", "Delete Story")
+        form.addTextInput("Type DELETE", "delete", "")
+
+        this.uiService.promptForm('', form).then((entries) => {
+          if (entries["delete"].value == "DELETE") {
+            this.storyService.deleteStory(this.storyId).then(() => {
+              this.router.navigate([''])
+            }).catch(e => {
+              this.alertService.error(e)
+            })
+          } else {
+            this.alertService.error("Invalid verification string")
+          }
         })
+        
       }).catch(e => {
         this.alertService.error(e)
       })
